refactor routing to have a centralized App component
